# sesi-08/swaggo/docs/swagger.yaml
basePath: /
definitions:
  main.Item:
    properties:
      description:
        example: A random description
        type: string
      itemId:
        example: A1B2C3
        type: string
      quantity:
        example: "1"
        type: string
    type: object
  main.Order:
    properties:
      customerName:
        type: string
      items:
        items:
          $ref: '#/definitions/main.Item'
        type: array
      orderId:
        type: string
      orderedAt:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
  description: This is a sample service for managing orders
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Orders API
  version: "1.0"
paths:
  /order/{orderId}:
    delete:
      consumes:
      - application/json
      description: Delete data order where orderId
      parameters:
      - description: ID
        in: path
        name: orderId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete data order where orderId
      tags:
      - orders
    get:
      consumes:
      - application/json
      description: Get details of all orders where orderId
      parameters:
      - description: ID
        in: path
        name: orderId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Order'
      summary: Get details of all orders where orderId
      tags:
      - orders
  /orders:
    get:
      consumes:
      - application/json
      description: Get details of all orders
      parameters:
      - description: Create order
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/main.Order'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Order'
      summary: Get details of all orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order with the input paylod
      parameters:
      - description: Create order
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/main.Order'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Order'
      summary: Create a new order
      tags:
      - orders
swagger: "2.0"